Added basic creation and deletion buttons for RFID
