+++ mhackspace/rfid/views.py
@@ -2,6 +2,10 @@ import logging
 from rest_framework.response import Response
 from rest_framework import viewsets
 from rest_framework import status
+
+from django.views.generic import ListView, DeleteView, CreateView
+from django.contrib.auth.mixins import LoginRequiredMixin
+
 from mhackspace.users.models import Rfid
 from mhackspace.rfid.models import Device, DeviceAuth
 from mhackspace.rfid.serializers import DeviceSerializer, AuthSerializer
@@ -39,3 +43,29 @@ class AuthUserWithDeviceViewSet(viewsets.ViewSet):
         serializer = AuthSerializer(
             instance={'name': device.name, 'rfid': rfid.code, 'device': device.identifier})
         return Response(serializer.data, status=200)
+
+
+class RfidCardsListView(LoginRequiredMixin, ListView):
+    template_name = 'users/rfid_form.html'
+    context_object_name = 'rfids'
+    paginate_by = 50
+    model = Rfid
+
+    def get_queryset(self):
+        return Rfid.objects.filter(user=self.request.user)
+
+
+class RfidCardsUpdateView(LoginRequiredMixin, CreateView):
+    fields = ['code', 'description', ]
+    model = Rfid
+    success_url = '/users/access-cards'
+
+    def form_valid(self, form):
+        user = self.request.user
+        form.instance.user = user
+        return super(RfidCardsUpdateView, self).form_valid(form)
+
+
+class RfidCardsDeleteView(LoginRequiredMixin, DeleteView):
+    model = Rfid
+    success_url = '/users/access-cards'
